# Use underlying slice in String() hex formatting for TraceID and SpanID