Fixes #480

Avoid the cast [16]byte and [8]byte for TraceID and SpanID
respectively when invoking their String methods and instead
take the underlying slice of each using [:] instead of
falling back to their arrays and then letting fmt.Sprintf
take their slices for us.

Empirical benchmarks produced
```shell
$ benchstat old.txt new.txt
name    old time/op    new time/op    delta
With-4     422ns ± 1%     187ns ± 1%  -55.60%          (p=0.008 n=5+5)

name    old alloc/op   new alloc/op   delta
With-4     64.0B ± 0%     64.0B ± 0%     ~     (all samples are equal)

name    old allocs/op  new allocs/op  delta
With-4      3.00 ± 0%      2.00 ± 0%  -33.33%          (p=0.008 n=5+5)
```

Final benchmarks still showed huge reductions in  ns/op
but discrepancies in throughput i.e bytes processed/op but overall
the same number of allocations so not a problem.

```shell
$ benchstat old.txt new.txt
name                 old time/op    new time/op    delta
TraceID_DotString-4     478ns ± 2%     260ns ± 1%  -45.60% (p=0.000 n=10+8)
SpanID_DotString-4      345ns ± 1%     225ns ± 2%  -34.83% (p=0.000 n=8+9)

name                 old alloc/op   new alloc/op   delta
TraceID_DotString-4     64.0B ± 0%     80.0B ± 0%  +25.00% (p=0.000 n=10+10)
SpanID_DotString-4      32.0B ± 0%     56.0B ± 0%  +75.00% (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
TraceID_DotString-4      3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
SpanID_DotString-4       3.00 ± 0%      3.00 ± 0%     ~     (all samples are equal)
```