@odeke-em 2018-02-23
Thank you for the review @rakyll, I've renamed the benchmark test variables, PTAL.