@odeke-em 2018-02-23
This change also improved SpanWithAnnotation benchmarks
```shell
SpanWithAnnotations_6-4     528ns ± 2%     512ns ± 1%   -2.99%          (p=0.029 n=4+4)
```
since Span.String() invokes SpanID.String() too.